created game and photos controllers
