+++ packages/backend/controllers/socketio/index.js
@@ -1,0 +1,2 @@
+//
+const onSocketConnect = () => {};
